backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6

pydantic-settings==2.5.2
python-dotenv==1.0.1
httpx==0.27.2
itsdangerous==2.2.0

sqlalchemy==2.0.34
psycopg2-binary==2.9.9

google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.144.0
openai==2.23.0